Make ADD_SPECIAL_TOKENS true by default
Before this change the ADD_SPECIAL_TOKENS acted as a tristate
variable where the default (not set) would be to add special
tokens only if the model didn't have a chat template.

After some deliberation where this default broke existing
integration tests, the decision was made to make the default
to be "true". Due to the way that environment variables work,
this means that this setting now only has 2 states: true and false.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>